Troquei a leitura de Epplus para ExcelMapper e decidi refazer a page, porém os codigos antigos estão comentados para vizualização!

+++ ImportExcel/Models/Autorizacao.cs
@@ -5,20 +5,19 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Web;
 
+
 namespace ImportExcel.Models
 {
     public class Autorizacao
     {
         [Key()]
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string? ObservacaoFuncionario { get; set; }
-        public int IdGerente1 { get; set; }
+        public long IdGerente1 { get; set; }
         public string? ObservacaoGerente1 { get; set; }
         public bool StatusGerente1 { get; set; }
-        public int IdGerente2 { get; set; }
+        public long IdGerente2 { get; set; }
         public string? ObservacaoGerente2 { get; set; }
         public bool StatusGerente2 { get; set; }
-
     }
-  
 }
